ET1/requirements.txt
# ET1: Semantic Annotation in Training Data
# Core dependencies

# ML / Training
torch>=2.1
transformers>=4.40
peft>=0.10
datasets>=2.19
accelerate>=0.30
bitsandbytes>=0.43

# Evaluation & Statistics
scikit-learn>=1.4
scipy>=1.12
numpy>=1.26

# Visualization
matplotlib>=3.8

# Data
pyyaml>=6.0

# jsonld-ex (for training data generation)
jsonld-ex>=0.7.1

# Testing
pytest>=8.0
pytest-cov>=5.0
